+++ modules/vcard/model/vcard.js
@@ -1,0 +1,16 @@
+/**
+ * Model for vCard
+ */
+'use strict';
+
+module.exports = {
+  get: () => {
+    return [
+      {
+        _version: '2.1',
+        name: 'Gerdes',
+        firstname: 'Uwe'
+      }
+    ]
+  }
+};